+++ js/ghu-install-vanilla.js
@@ -1,79 +1,88 @@
 /**
- * Javascript to show and hide the API specific settings
+ * Vanilla Javascript to show and hide the API specific settings
  * for the remote install feature.
  *
  * @class  Fragen\GitHub_Updater\Install
- * @since  4.6.0
+ * @since  8.5.0
  * @access public
  * @package	github-updater
- *
- * Was working, then stopped working.
- * Seems not to get all 3 bitbucket_setting elements.
  */
 
 (function () {
 
 	// Hide non-default (Bitbucket & GitLab) settings on page load.
-	var nonDefault = ['bitbucket', 'gitlab', 'gitea'];
+	var nonDefault = ['bitbucket', 'gitlab', 'gitea', 'zipfile'];
 
-	nonDefault.forEach(
-		function (item) {
-			var parents              = getParents( item, 'tr' );
-			parents[0].style.display = 'none';
-		}
-	);
+	nonDefault.forEach(function (item) {
+		var parents = getParents(item, 'tr');
+		displayNone(parents);
+	});
 
 	// When the api selector changes.
-	var selects = document.querySelector( 'select[ name="github_updater_api" ]' );
+	var selects = document.querySelector('select[ name="github_updater_api" ]');
 
-	selects.addEventListener(
-		'change',
-		function () {
-			var defaults = ['github', 'bitbucket', 'gitlab', 'gitea'];
+	// Only run when on proper tab.
+	if (selects !== null) {
+		selects.addEventListener('change', function () {
+			var defaults = ['github', 'bitbucket', 'gitlab', 'gitea', 'zipfile'];
 
 			// Create difference array.
-			var hideMe = remove( defaults, this.value );
+			var hideMe = remove(defaults, this.value);
 
 			// Hide items with unselected api's classes.
-			hideMe.forEach(
-				function (item) {
-					var parents              = getParents( item, 'tr' );
-					parents[0].style.display = 'none';
-				}
-			);
+			hideMe.forEach(function (item) {
+				var parents = getParents(item, 'tr');
+				displayNone(parents);
+			});
 
 			// Show selected setting.
-			[this.value].forEach(
-				function (item) {
-					var parents              = getParents( item, 'tr' );
-					parents[0].style.display = '';
-				}
-			);
-		}
-	);
+			[this.value].forEach(function (item) {
+				var parents = getParents(item, 'tr');
+				display(parents);
+			});
+		});
+	}
 
 	// Remove selected element from array and return array.
 	function remove(array, element) {
-		const index = array.indexOf( element );
+		const index = array.indexOf(element);
 		if (index !== -1) {
-			array.splice( index, 1 );
+			array.splice(index, 1);
 		}
-
 		return array;
+	}
+
+	// Hide element.
+	function displayNone(array) {
+		array.forEach((item) => {
+			item.style.display = 'none';
+		});
+	}
+
+	// Display element.
+	function display(array) {
+		array.forEach((item) => {
+			item.style.display = '';
+		});
 	}
 
 	// Return query and selector for `$(query).parents.(selector)`.
 	function getParents(item, selector) {
-		return vanillaParents( document.querySelector( 'input.'.concat( item, '_setting' ) ), selector );
+		return vanillaParents(document.querySelectorAll('input.'.concat(item, '_setting')), selector);
 	}
 
 	// Vanilla JS version of jQuery `$(query).parents(selector)`.
 	function vanillaParents(element, selector) {
 		var parents = [];
-		while (element = element.parentElement.closest( selector )) {
-			parents.push( element );
+		if (NodeList.prototype.isPrototypeOf(element)) {
+			element.forEach((item) => {
+				element = item.parentElement.closest(selector);
+				parents.push(element);
+			});
+		} else {
+			element = item.parentElement.closest(selector);
+			parents.push(element);
 		}
-
 		return parents;
 	}