+++ src/GitHub_Updater/API/GitLab_API.php
@@ -181,10 +181,10 @@ class GitLab_API extends API implements API_Interface {
 		/*
 		 * If release asset.
 		 */
-		if ( $this->type->release_asset && '0.0.0' !== $this->type->newest_tag ) {
-			$download_link_base = $this->make_release_asset_download_link();
-
-			return $this->add_access_token_endpoint( $this, $download_link_base );
+		if ( $this->type->ci_job && '0.0.0' !== $this->type->newest_tag ) {
+			$release_asset = $this->get_release_asset();
+			$release_asset = $this->get_release_asset_redirect( $release_asset );
+			return $this->add_access_token_endpoint( $this, $release_asset );
 		}
 
 		/*
@@ -213,27 +213,25 @@ class GitLab_API extends API implements API_Interface {
 	}
 
 	/**
-	 * Create release asset download link.
-	 * Filename must be `{$slug}-{$newest_tag}.zip`
-	 *
-	 * @access private
+	 * Get GitLab release asset download link.
 	 *
 	 * @return string $download_link
 	 */
-	private function make_release_asset_download_link() {
-		$download_link = implode(
-			'/',
-			[
-				'https://gitlab.com/api/v4/projects',
-				urlencode( $this->type->owner . '/' . $this->type->slug ),
-				'builds/artifacts',
-				$this->type->newest_tag,
-				'download',
-			]
-		);
-		$download_link = add_query_arg( 'job', $this->type->ci_job, $download_link );
+	public function get_release_asset() {
+		self::$method = 'release_asset';
+		$response     = isset( $this->response['release_asset'] ) ? $this->response['release_asset'] : false;
 
-		return $download_link;
+		if ( $response && $this->exit_no_update( $response ) ) {
+			return false;
+		}
+
+		if ( ! $response ) {
+			$response = $this->get_api_url( "/:owner/:repo/-/jobs/artifacts/{$this->type->newest_tag}/download", true );
+			$response = add_query_arg( 'job', $this->type->ci_job, $response );
+			$this->set_repo_cache( 'release_asset', $response );
+		}
+
+		return $response;
 	}
 
 	/**
@@ -252,6 +250,7 @@ class GitLab_API extends API implements API_Interface {
 			case 'meta':
 			case 'tags':
 			case 'branches':
+			case 'release_asset':
 				break;
 			case 'file':
 			case 'changes':