+++ src/GitHub_Updater/Init.php
@@ -55,8 +55,8 @@ class Init extends Base {
 			update_site_option( 'github_updater', array_merge( $options, [ 'current_branch_github-updater' => 'develop' ] ) );
 		}
 
-		if ( 'github-updater/github-updater.php' !== $slug ) {
-			rename( $plugin_dir . dirname( $slug ), $plugin_dir . 'github-updater' );
+		if ( $slug && 'github-updater/github-updater.php' !== $slug ) {
+			@rename( $plugin_dir . dirname( $slug ), $plugin_dir . 'github-updater' );
 		}
 	}