+++ CHANGES.md
@@ -1,5 +1,16 @@
 #### [unreleased]
 
+#### 8.5.0 / 2018-11-26
+* silence rename PHP warning during plugin update
+* specify branch for changelog
+* refactored dot org override, constant deprecated in favor of new filter `github_updater_override_dot_org`
+* now using vanilla JS for Install settings
+* refactored GitHub release asset code to get direct download link
+* refactored Bitbucket release asset code to get redirected download link for AWS
+* refactored GitLab release asset code to get redirected download link
+* exit early if checking _View details_ but not done with background update, avoids PHP notices
+* updated to add/use composer dependencies and autoloader
+
 #### 8.4.2 / 2018-11-01
 * updated password fields to not autoload saved passwords, thanks @figureone
 * fixed error when saving Remote Management options